CORE-285 Moving commit_upgrade_if_needed to a separate file for more clarity
