+++ node/src/reactor/main_reactor/control.rs
@@ -1,18 +1,14 @@
-use std::{collections::BTreeMap, sync::Arc, time::Duration};
-use tokio::runtime::Handle;
+use std::{collections::BTreeMap, time::Duration};
 use tracing::{debug, error, info, trace};
 
 use casper_storage::data_access_layer::GenesisResult;
-use casper_types::{
-    BlockHash, BlockHeader, Chainspec, ChainspecRawBytes, Digest, EraId, PublicKey, TimeDiff,
-    Timestamp,
-};
+use casper_types::{BlockHash, BlockHeader, Digest, EraId, PublicKey, Timestamp};
 
 use crate::{
     components::{
         binary_port,
         block_synchronizer::{self, BlockSynchronizerProgress},
-        contract_runtime::{ContractRuntime, ExecutionPreState},
+        contract_runtime::ExecutionPreState,
         diagnostics_port, event_stream_server, network, rest_server, storage, upgrade_watcher,
     },
     effect::{announcements::ControlAnnouncement, EffectBuilder, EffectExt, Effects},
@@ -20,8 +16,7 @@ use crate::{
     reactor::main_reactor::{
         catch_up::CatchUpInstruction, genesis_instruction::GenesisInstruction,
         keep_up::KeepUpInstruction, upgrade_shutdown::UpgradeShutdownInstruction, utils,
-        validate::ValidateInstruction, Error, MainEvent, MainReactor, PendingImmediateSwitchBlock,
-        ReactorState,
+        validate::ValidateInstruction, MainEvent, MainReactor, ReactorState,
     },
     types::{BlockPayload, ExecutableBlock, FinalizedBlock, InternalEraReport, MetaBlockState},
     NodeRng,
@@ -430,93 +425,6 @@ impl MainReactor {
             // non-validators should start receiving gossip about the block at height 1 soon
             GenesisInstruction::NonValidator(self.control_logic_default_delay.into(), effects)
         }
-    }
-
-    /// If `tip_header` is a switch block that is the last block before the chainspec's
-    /// activation point, synchronously commits the protocol upgrade against `contract_runtime`'s
-    /// global state. Returns the info needed to later produce, sign, and gossip the resulting
-    /// immediate switch block, once the reactor is ready to do so (see
-    /// [`Self::maybe_finish_pending_upgrade`]). Returns `Ok(None)` if no upgrade is due.
-    ///
-    /// This is an associated function (rather than a `&self` method) so it can be called from
-    /// `MainReactor::new`, before the reactor itself has been constructed -- that's the only
-    /// call site: a fresh restart whose local tip already sits at the pre-activation switch
-    /// block (e.g. after a live node shuts itself down for the upgrade). A node still *catching
-    /// up* through a historical activation point does not go through here; it just receives the
-    /// post-upgrade chain via the ordinary block-synchronizer fetch path, like any other
-    /// historical data.
-    pub(super) fn commit_upgrade_if_needed(
-        contract_runtime: &ContractRuntime,
-        chainspec: &Arc<Chainspec>,
-        chainspec_raw_bytes: &Arc<ChainspecRawBytes>,
-        tip_header: Option<&BlockHeader>,
-        upgrade_timeout: TimeDiff,
-    ) -> Result<Option<PendingImmediateSwitchBlock>, Error> {
-        let Some(tip_header) = tip_header else {
-            return Ok(None);
-        };
-        if !(tip_header.is_switch_block()
-            && tip_header.is_last_block_before_activation(&chainspec.protocol_config))
-        {
-            return Ok(None);
-        }
-
-        info!(
-            era_id = %tip_header.era_id(),
-            height = tip_header.height(),
-            "committing protocol upgrade"
-        );
-
-        let upgrade_config = chainspec
-            .upgrade_config_from_parts(
-                *tip_header.state_root_hash(),
-                tip_header.protocol_version(),
-                chainspec.protocol_config.activation_point.era_id(),
-                chainspec_raw_bytes.clone(),
-            )
-            .map_err(Error::ProtocolUpgrade)?;
-
-        // Executing protocol upgrade can be time consuming. It's executed in the background so the
-        // upgrade_timeout can be enforced. This function stays synchronous -- it's called
-        // from `MainReactor::new`, before the reactor's async event loop exists -- so the
-        // wait for that bounded future to resolve is bridged onto a dedicated scoped
-        // thread, which calls `Handle::block_on` directly.
-        let handle = Handle::current();
-        let post_state_hash = std::thread::scope(|scope| {
-            scope
-                .spawn(|| {
-                    handle.block_on(async {
-                        match tokio::time::timeout(
-                            Duration::from(upgrade_timeout),
-                            contract_runtime.commit_protocol_upgrade(upgrade_config),
-                        )
-                        .await
-                        {
-                            Ok(result) => result,
-                            Err(_) => Err(format!(
-                                "protocol upgrade did not complete within {}",
-                                upgrade_timeout
-                            )),
-                        }
-                    })
-                })
-                .join()
-                .unwrap_or_else(|panic| std::panic::resume_unwind(panic))
-        })
-        .map_err(Error::ProtocolUpgrade)?;
-
-        Ok(Some(PendingImmediateSwitchBlock {
-            next_block_height: tip_header.height() + 1,
-            post_state_hash,
-            parent_hash: tip_header.block_hash(),
-            parent_seed: *tip_header.accumulated_seed(),
-            era_id: tip_header.next_block_era_id(),
-            // Adding one second here to make sure the timestamp is monotonically growing -
-            // it's important for EVM smart contracts
-            timestamp: tip_header
-                .timestamp()
-                .saturating_add(TimeDiff::from_seconds(1)),
-        }))
     }
 
     /// If a protocol upgrade has been committed and its immediate switch block hasn't yet been
